portal_f1/requirements.txt
psycopg2-binary
asgiref==3.5.2
autopep8==1.6.0
django==4.0.5
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
Unipath==1.1
dj-database-url==0.5.0
python-decouple==3.6
gunicorn==20.1.0
whitenoise==6.2.0
django-braces==1.15.0
django-cleanup==6.0.0
djangorestframework==3.13.1
django-filter==22.1
django-localflavor==3.1
networkx==2.8.4

djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
hurry.filesize==0.9
django-recaptcha==3.0.0
django-rest-swagger==2.2.0
django-filter==22.1

pillow==9.5.0
django-imagekit==4.1.0

python-dotenv==1.0.0
